Add modifications and new components for placing orders
