+++ src/components/Banner/Banner.scss
@@ -59,12 +59,14 @@
   font-family: Pragmatica;
   font-weight: 500;
   font-size: 16px;
+  padding-right: 5px;
 }
 
 .banner-price {
   font-family: Pragmatica;
   font-weight: 500;
   font-size: 22px;
+  padding-right: 5px;
 }
 
 .uah {
@@ -99,7 +101,7 @@
   display: none;
 }
 
-@media screen and (min-width: 1200px) and (max-width: 1640px) {
+@media screen and (min-width: 1280px) and (max-width: 1640px) {
   .banner-timer {
     display: block;
   }
@@ -164,7 +166,7 @@
   }
 }
 
-@media screen and (min-width: 768px) and (max-width: 1200px) {
+@media screen and (min-width: 768px) and (max-width: 1280px) {
   .banner-timer {
     display: block;
   }
