+++ src/components/Button/SliderCircle/SliderCircle.jsx
@@ -5,16 +5,16 @@ import "../SliderCircle/SliderCircle.scss";
 function SliderCircle() {
   return (
     <div className="SliderCircle">
-      <button className="SliderCircleLeft">
+      <button className="sliderCircleLeft">
         <ul className="buttonComponents">
-          <li className="SliderCircle-icons">
+          <li className="sliderCircle-icons">
             <img src={ArrowLeft} alt="ArrowLeft" />
           </li>
         </ul>
       </button>
-      <button className="SliderCircleRight">
+      <button className="sliderCircleRight">
         <ul className="buttonComponents">
-          <li className="SliderCircle-icons">
+          <li className="sliderCircle-icons">
             <img src={ArrowRight} alt="ArrowRight" />
           </li>
         </ul>
